feat: fix webhooks registration in context of settings saving (#3)

+++ Observer/AdminWebhookRegister.php
@@ -32,21 +32,26 @@ class AdminWebhookRegister implements ObserverInterface {
 
     public function execute(\Magento\Framework\Event\Observer $observer)
     {
-        $body = $this->client->retrieveWebhooks();
-        $webhooks = json_decode($body, true);
+        $mode = $this->config->getConfigData('paymaya_mode', 'basic');
+        $encryptedSecretKey = $this->config->getConfigData("paymaya_{$mode}_sk", 'basic');
 
-        foreach ($webhooks as $webhook) {
-            if (in_array($webhook['name'], $this->overridable_webhooks)) {
-                $this->client->deleteWebhook($webhook['id']);
+        if ($this->config->isEnabled() && isset($encryptedSecretKey)) {
+            $body = $this->client->retrieveWebhooks();
+            $webhooks = json_decode($body, true);
+
+            foreach ($webhooks as $webhook) {
+                if (in_array($webhook['name'], $this->overridable_webhooks)) {
+                    $this->client->deleteWebhook($webhook['id']);
+                }
             }
+
+            $webhookBaseUrl = $this->config->getConfigData('webhook_base_url', 'webhooks');
+
+            $this->client->createWebhook('CHECKOUT_SUCCESS', "{$webhookBaseUrl}/paymaya/webhooks");
+            $this->client->createWebhook('CHECKOUT_FAILURE', "{$webhookBaseUrl}/paymaya/webhooks");
+            $this->client->createWebhook('PAYMENT_SUCCESS', "{$webhookBaseUrl}/paymaya/webhooks/payment");
+            $this->client->createWebhook('PAYMENT_FAILED', "{$webhookBaseUrl}/paymaya/webhooks/payment");
+            $this->client->createWebhook('PAYMENT_EXPIRED', "{$webhookBaseUrl}/paymaya/webhooks/payment");
         }
-
-        $webhookBaseUrl = $this->config->getConfigData('webhook_base_url', 'webhooks');
-
-        $this->client->createWebhook('CHECKOUT_SUCCESS', "{$webhookBaseUrl}/paymaya/webhooks");
-        $this->client->createWebhook('CHECKOUT_FAILURE', "{$webhookBaseUrl}/paymaya/webhooks");
-        $this->client->createWebhook('PAYMENT_SUCCESS', "{$webhookBaseUrl}/paymaya/webhooks/payment");
-        $this->client->createWebhook('PAYMENT_FAILED', "{$webhookBaseUrl}/paymaya/webhooks/payment");
-        $this->client->createWebhook('PAYMENT_EXPIRED', "{$webhookBaseUrl}/paymaya/webhooks/payment");
     }
 }
